+++ kaithem/src/thirdparty/iot_devices/devices/RTL433/__init__.py
@@ -15,11 +15,10 @@ import asyncio
 
 eventLoop = asyncio.new_event_loop()
 
-
 mqttlock = threading.Lock()
 
-
 import iot_devices.device as devices
+
 
 @asyncio.coroutine
 def scan():
@@ -32,12 +31,12 @@ def scan():
                     # it gone, because packet loss will be less
                     m = 3 if all_devs.datapoints['rssi'] > -65 else 7
 
-                    if all_devs.lastseen < time.monotonic() - (float(all_devs.config.get('interval',60) or 60) * m):
+                    if all_devs.lastseen < time.monotonic() - (float(
+                            all_devs.config.get('interval', 60) or 60) * m):
                         # This is how we mark it as not there
                         all_devs.set_data_point('rssi', -180)
             except:
                 logging.exception("RTL err")
-
 
 
 t = threading.Thread(target=eventLoop.run_forever, name="RTL433Task")
@@ -48,7 +47,6 @@ eventLoop.call_soon_threadsafe(scan)
 from mako.lookup import TemplateLookup
 
 templateGetter = TemplateLookup(os.path.dirname(__file__))
-
 
 defaultSubclassCode = """
 class CustomDeviceType(DeviceType):
@@ -67,29 +65,31 @@ class RTL433Client(devices.Device):
 
     def onConnectionChange(self, status):
         if status == "connected":
-            self.set_data_point("mqttStatus",1)
+            self.set_data_point("mqttStatus", 1)
         else:
-            self.set_data_point("mqttStatus",0)
-
-
+            self.set_data_point("mqttStatus", 0)
 
     def __init__(self, name, data):
         devices.Device.__init__(self, name, data)
 
         try:
-            self.set_config_default('device.interval','300')
+            self.set_config_default('device.interval', '300')
 
-            self.numeric_data_point("rssi", min=-180, 
-                max=12, interval=float(self.config["device.interval"]),
-                 description="-75 if recetly seen, otherwise -180, we don't have real RSSI data",
-                 writable=False)
+            self.numeric_data_point(
+                "rssi",
+                min=-180,
+                max=12,
+                interval=float(self.config["device.interval"]),
+                description=
+                "-75 if recetly seen, otherwise -180, we don't have real RSSI data",
+                writable=False)
 
-            self.set_config_default('device.id','')
-            self.set_config_default('device.model','')
-            self.set_config_default('device.server','localhost')
-            self.set_config_default('device.port','1883')
-            self.set_config_default('device.password','')
-            self.set_config_default('device.mqttTopic','home/rtl_433')
+            self.set_config_default('device.id', '')
+            self.set_config_default('device.model', '')
+            self.set_config_default('device.server', 'localhost')
+            self.set_config_default('device.port', '1883')
+            self.set_config_default('device.password', '')
+            self.set_config_default('device.mqttTopic', 'home/rtl_433')
 
             # This connection is actually  possibly shared
             # Scullery does the deduplication for us
@@ -99,12 +99,12 @@ class RTL433Client(devices.Device):
                 self.config["device.server"],
                 int(self.config["device.port"].strip() or 1883),
                 password=self.config["device.password"].strip(),
-                connectionID=str("RTL433Connection")
-            )
-           
-            self.numeric_data_point("mqttStatus",writable=False)
+                connectionID=str("RTL433Connection"))
+
+            self.numeric_data_point("mqttStatus", writable=False)
             self.connection.subscribeToStatus(self.onConnectionChange)
-            self.set_data_point("mqttStatus", 1 if self.connection.isConnected else 0)
+            self.set_data_point("mqttStatus",
+                                1 if self.connection.isConnected else 0)
 
             topic = data.get("device.mqtttopic", "home/rtl_433")
 
@@ -113,93 +113,113 @@ class RTL433Client(devices.Device):
             def onBattery(t, m):
                 m = float(m)
                 if not 'battery' in self.datapoints:
-                    self.numeric_data_point("battery",default=50,writable=False,unit="%")
+                    self.numeric_data_point("battery",
+                                            default=50,
+                                            writable=False,
+                                            unit="%")
 
                     # Always set before setting the alarm.
                     self.set_data_point("battery", m)
 
-                    self.set_alarm(
-                        name="Low battery", datapoint="wind", expression="value < 15", priority="info")
+                    self.set_alarm(name="Low battery",
+                                   datapoint="battery",
+                                   expression="value < 15",
+                                   priority="info")
 
                 self.set_data_point("battery", m)
 
             def onWind(t, m):
                 m = float(m)
                 if not 'wind' in self.datapoints:
-                    self.numeric_data_point("wind",unit = "km/h",writable=False)
+                    self.numeric_data_point("wind",
+                                            unit="km/h",
+                                            writable=False)
 
-                    self.set_alarm(
-                        name="High Wind", datapoint="wind", expression="value > 35", priority="info")
+                    self.set_alarm(name="High Wind",
+                                   datapoint="wind",
+                                   expression="value > 35",
+                                   priority="info")
 
                 self.set_data_point("wind", m)
 
             def onTemp(t, m):
                 m = float(m)
                 if not 'temp' in self.datapoints:
-                    self.numeric_data_point("temp",unit = "degC",writable=False)
-                    self.set_alarm(
-                        name="Freezing temperatures", datapoint="wind", expression="value < 2", priority="info")
+                    self.numeric_data_point("temp",
+                                            unit="degC",
+                                            writable=False)
+                    self.set_alarm(name="Freezing temperatures",
+                                   datapoint="temp",
+                                   expression="value < 2",
+                                   priority="info")
 
                 self.set_data_point("temp", m)
 
             def onHum(t, m):
                 m = float(m)
                 if not 'humidity' in self.datapoints:
-                    self.numeric_data_point("humidity",unit = "%",writable=False)
-                    self.set_alarm(
-                        name="High humidity", datapoint="humidity", expression="value > 80", priority="info")
-                    self.set_alarm(
-                        name="Low humidity", datapoint="humidity", expression="value < 20", priority="info")
+                    self.numeric_data_point("humidity",
+                                            unit="%",
+                                            writable=False)
+                    self.set_alarm(name="High humidity",
+                                   datapoint="humidity",
+                                   expression="value > 80",
+                                   priority="info")
+                    self.set_alarm(name="Low humidity",
+                                   datapoint="humidity",
+                                   expression="value < 20",
+                                   priority="info")
 
                 self.set_data_point("humidity", m)
-
 
             def onMoist(t, m):
                 m = float(m)
                 if not 'moisture' in self.datapoints:
-                    self.numeric_data_point("moisture",unit = "%",writable=False)
+                    self.numeric_data_point("moisture",
+                                            unit="%",
+                                            writable=False)
                 self.set_data_point("moisture", m)
-
 
             def onPres(t, m):
                 m = float(m)
                 if not 'pressure' in self.datapoints:
-                    self.numeric_data_point("pressure",unit = "Pa",writable=False)
+                    self.numeric_data_point("pressure",
+                                            unit="Pa",
+                                            writable=False)
                 self.set_data_point("pressure", m)
-
 
             def onWeight(t, m):
                 m = float(m)
                 if not 'weight' in self.datapoints:
-                    self.numeric_data_point("weight",writable=False)
+                    self.numeric_data_point("weight", writable=False)
                 self.set_data_point("weight", m)
-
 
             def onCommandCode(t, m):
                 m = float(m)
                 if not 'lastCommandCode' in self.datapoints:
-                    self.object_data_point("lastCommandCode",writable=False)
-                self.set_data_point("lastCommandCode", (m,time.time()))
+                    self.object_data_point("lastCommandCode", writable=False)
+                self.set_data_point("lastCommandCode", (m, time.time()))
 
             def onCommandName(t, m):
                 m = float(m)
                 if not 'lastCommandName' in self.datapoints:
-                    self.object_data_point("lastCommandName",writable=False)
-                self.set_data_point("lastCommandName", (m,time.time()))
+                    self.object_data_point("lastCommandName", writable=False)
+                self.set_data_point("lastCommandName", (m, time.time()))
 
-     
             def onJSON(t, m):
                 m = json.loads(m)
                 self.print(m, "Saw packet on air")
 
                 # Going to do an ID match.
                 if 'device.id' in self.config and self.config['device.id']:
-                    if not ('id' in m and str(m['id']) == self.config['device.id']):
+                    if not ('id' in m
+                            and str(m['id']) == self.config['device.id']):
                         self.print(m, "Packet filter miss")
                         return
 
                 if 'device.model' in self.config and self.config['device.id']:
-                    if not ('model' in m and m['model'] == self.config['device.model']):
+                    if not ('model' in m
+                            and m['model'] == self.config['device.model']):
                         self.print(m, "Packet filter miss")
                         return
 
@@ -233,33 +253,30 @@ class RTL433Client(devices.Device):
 
                 if 'cmd' in m:
                     onCommandCode(0, m['cmd'])
-                
 
                 if 'button_id' in m:
                     onCommandCode(0, m['button_id'])
-
 
                 if 'button_name' in m:
                     onCommandName(0, m['button_name'])
 
                 if 'event' in m:
                     onCommandName(0, m['event'])
-                
+
                 if 'code' in m:
                     onCommandCode(0, m['code'])
 
             self.noGarbage = [onJSON]
 
-            self.connection.subscribe(
-                    topic, onJSON, encoding="raw")
+            self.connection.subscribe(topic, onJSON, encoding="raw")
 
-            all_devs[self.name]=self
-        except:
+            all_devs[self.name] = self
+        except Exception as e:
             self.handle_exception()
+
+    def close(self):
+        return super().close()
+        self.connection.unsubscribe(self.noGarbage[0])
 
     def getManagementForm(self):
         return ""
-        #return templateGetter.get_template("manageform.html").render(data=self.config, obj=self)
-
-
-