+++ kaithem/src/docs/changes.md
@@ -5,7 +5,9 @@ Change Log
 ### 0.67.6
 
 - Semi-breaking: Tag point alarms will not trigger if the tag point has never actually had a value set.
- 
+- Support for searching all modules for cross-framework devices, and importing on demand.
+- Fix devices in modules bugs
+
 ### 0.67.5
 - Scheduler is now just based on the normal sched module
 - Various performance improvments(Seems like 50% les CPU usage!)